+++ src/elements/PaymentForm.php
@@ -14,10 +14,8 @@ use craft\behaviors\FieldLayoutBehavior;
 use craft\elements\db\ElementQueryInterface;
 use craft\helpers\Json;
 use enupal\stripe\models\Settings;
-use enupal\stripe\enums\DiscountType;
 use enupal\stripe\enums\SubscriptionType;
 use enupal\stripe\Stripe;
-use enupal\stripe\validators\DiscountValidator;
 use craft\helpers\UrlHelper;
 use craft\elements\actions\Delete;
 
@@ -85,8 +83,6 @@ class PaymentForm extends Element
     public $hasUnlimitedStock;
     public $customerQuantity;
     public $soldOutMessage;
-    public $discountType;
-    public $discount;
     public $shippingAmount;
     public $itemWeight;
     public $itemWeightUnit;
@@ -205,62 +201,6 @@ class PaymentForm extends Element
         $tax = $this->settings->tax ?? null;
 
         return $tax;
-    }
-
-    /**
-     * @return string
-     */
-    public function getTaxType()
-    {
-        $taxType = null;
-
-        switch ($this->settings->taxType) {
-            case DiscountType::RATE:
-                {
-                    $taxType = 'tax_rate';
-                    break;
-                }
-            case DiscountType::AMOUNT:
-                {
-                    $taxType = 'rate';
-                    break;
-                }
-        }
-
-        return $taxType;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDiscount()
-    {
-        $discount = $this->discount ?? null;
-
-        return $discount;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDiscountType()
-    {
-        $discountType = null;
-
-        switch ($this->discountType) {
-            case DiscountType::RATE:
-                {
-                    $discountType = 'discount_rate';
-                    break;
-                }
-            case DiscountType::AMOUNT:
-                {
-                    $discountType = 'discount_amount';
-                    break;
-                }
-        }
-
-        return $discountType;
     }
 
     /**
@@ -512,8 +452,6 @@ class PaymentForm extends Element
         $record->enableRememberMe = $this->enableRememberMe;
         $record->quantity = $this->quantity;
         $record->hasUnlimitedStock = $this->hasUnlimitedStock;
-        $record->discountType = $this->discountType;
-        $record->discount = $this->discount;
 
         $record->verifyZip = $this->verifyZip;
         $record->enableBillingAddress = $this->enableBillingAddress;
@@ -567,11 +505,7 @@ class PaymentForm extends Element
                 return $model->enableCheckout != 1;
             }],
             [['name', 'handle'], 'string', 'max' => 255],
-            [['name', 'handle'], UniqueValidator::class, 'targetClass' => PaymentFormRecord::class],
-            [
-                ['discount'],
-                DiscountValidator::class
-            ],
+            [['name', 'handle'], UniqueValidator::class, 'targetClass' => PaymentFormRecord::class]
         ];
     }
 
@@ -596,7 +530,11 @@ class PaymentForm extends Element
      * @param array|null $options
      *
      * @return string
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\RuntimeError
+     * @throws \Twig\Error\SyntaxError
      * @throws \yii\base\Exception
+     * @throws \yii\base\InvalidConfigException
      */
     public function paymentForm(array $options = null)
     {
@@ -615,6 +553,7 @@ class PaymentForm extends Element
         $logoUrl = null;
         $logoAsset = $this->getLogoAsset();
         $calculateFinalAmount = $options['calculateFinalAmount'] ?? true;
+        $couponData = $this->getCouponData($options);
 
         if ($logoAsset){
             $logoUrl = $logoAsset->getUrl();
@@ -696,6 +635,7 @@ class PaymentForm extends Element
             'paymentTypeIds' => $paymentTypeIds,
             'enableShippingAddress' => $this->enableShippingAddress,
             'enableBillingAddress' => $this->enableBillingAddress,
+            'coupon' => $couponData,
             'stripe' => [
                 'description' => $this->name,
                 'panelLabel' =>  $this->checkoutButtonText ?? 'Pay {{amount}}',
@@ -749,6 +689,8 @@ class PaymentForm extends Element
      * @param string $default
      *
      * @return string
+     * @throws \Twig\Error\LoaderError
+     * @throws \Twig\Error\SyntaxError
      */
     public function getPaymentFormText($default = null)
     {
@@ -856,5 +798,23 @@ class PaymentForm extends Element
         $methods = json_decode($this->paymentType, true);
 
         return $methods[0] ?? '';
+    }
+
+    /**
+     * @param $options
+     * @return array
+     */
+    private function getCouponData($options)
+    {
+        $couponData = [
+            'enabled' => $options['coupon']['enabled'] ?? false,
+            'displayTotal' => $options['coupon']['displayTotal'] ?? false,
+            'totalAmountLabel' => $options['coupon']['totalAmountLabel'] ?? false,
+            'label' => $options['coupon']['label'] ?? Craft::t('site', 'Coupon Code'),
+            'successMessage' => $options['coupon']['successMessage'] ?? '{name} - {id}',
+            'errorMessage' => $options['coupon']['errorMessage'] ?? Craft::t('site','This coupon is not valid'),
+        ];
+
+        return $couponData;
     }
 }