+++ src/resources/stripe/js/enupal-stripe-elements.js
@@ -57,6 +57,8 @@ var enupalStripe = {};
             // get the form ID
             var enupalStripeData = $.parseJSON($(enupalButtonElement).find('[name="enupalStripe[stripeData]"]').val());
 
+            var that = this;
+
             //  Firefox is cached for some reason when we empty the hidden input.
             if (navigator.userAgent.indexOf("Firefox") < 0) {
                 // reset our values
@@ -114,6 +116,23 @@ var enupalStripe = {};
                 });
             }
 
+            if (enupalStripeData.coupon.enabled){
+                var $couponButton = $("#check-coupon-button-"+enupalStripeData.paymentFormId);
+                this.updateTotalAmoutLabel(enupalButtonElement, enupalStripeData);
+                var $removeCoupon = enupalButtonElement.find("#remove-coupon-"+enupalStripeData.paymentFormId);
+                this.updatesTotalLabelOnChoosesPlan(enupalButtonElement, enupalStripeData);
+
+                $removeCoupon.click(function(event) {
+                    event.preventDefault();
+                    that.handleRemoveCoupon(enupalButtonElement, enupalStripeData, that);
+                });
+
+                $couponButton.click(function(event) {
+                    event.preventDefault();
+                    that.handleCouponValidation(enupalButtonElement, enupalStripeData, that);
+                });
+            }
+
             for ( var i = 0, l = pTypes.length; i < l; i++ ) {
                 if (pTypes[i] == 1){// Credit Card
                     if (i == 0){
@@ -138,6 +157,84 @@ var enupalStripe = {};
                 }
             }
 
+        },
+
+        updateTotalAmoutLabel(enupalButtonElement, enupalStripeData)
+        {
+            var $totalMessage = enupalButtonElement.find("#total-amount-value-"+enupalStripeData.paymentFormId);
+            var amount = this.getFinalAmount(enupalButtonElement, enupalStripeData);
+            if ($totalMessage){
+                $totalMessage.text(amount);
+            }
+        },
+
+        handleRemoveCoupon(enupalButtonElement, enupalStripeData, that)
+        {
+            var $couponMessage = enupalButtonElement.find("#coupon-message-"+enupalStripeData.paymentFormId);
+            var $removeCoupon = enupalButtonElement.find("#remove-coupon-"+enupalStripeData.paymentFormId);
+            var $couponInput = enupalButtonElement.find("#couponCode-"+enupalStripeData.paymentFormId);
+            if ($couponInput){
+                $couponInput.val('');
+            }
+            if ($couponMessage){
+                $couponMessage.text('');
+            }
+            $removeCoupon.addClass("hidden");
+
+            enupalButtonElement.find('[name="enupalCouponCode"]').val('');
+            that.updateTotalAmoutLabel(enupalButtonElement, enupalStripeData);
+        },
+
+        handleCouponValidation(enupalButtonElement, enupalStripeData, that)
+        {
+            var $couponButton = enupalButtonElement.find("#check-coupon-button-"+enupalStripeData.paymentFormId);
+            var $couponMessage = enupalButtonElement.find("#coupon-message-"+enupalStripeData.paymentFormId);
+            var $removeCoupon = enupalButtonElement.find("#remove-coupon-"+enupalStripeData.paymentFormId);
+            var $totalMessage = enupalButtonElement.find("#total-amount-value-"+enupalStripeData.paymentFormId);
+            var $couponInput = enupalButtonElement.find("#couponCode-"+enupalStripeData.paymentFormId);
+            var couponCode = $couponInput.val();
+            var stripeConfig = enupalStripeData.stripe;
+            $couponInput.val('');
+
+            var amount = this.convertToCents(that.getFinalAmount(enupalButtonElement, enupalStripeData), stripeConfig.currency);
+            $couponButton.prop('disabled', true);
+            var isRecurring = this.getIsRecurring(enupalButtonElement, enupalStripeData);
+
+            var data = {
+                'amount' : amount,
+                'couponCode': couponCode,
+                'isRecurring': isRecurring,
+                'currency': stripeConfig.currency,
+                'successMessage': enupalStripeData.coupon.successMessage
+            };
+
+            $.ajax({
+                type:"POST",
+                url:"enupal/validate-coupon",
+                data: data,
+                dataType : 'json',
+                success: function(response) {
+                    if (response.success === true){
+                        $couponMessage.removeClass('coupon-error');
+                        $couponMessage.text(response.successMessage);
+                        $removeCoupon.removeClass('hidden');
+
+                        if ($totalMessage){
+                            $totalMessage.text(response.finalAmount);
+                        }
+                        enupalButtonElement.find('[name="enupalCouponCode"]').val(response.coupon.id);
+                    }else{
+                        $couponMessage.addClass('coupon-error');
+                        $removeCoupon.addClass('hidden');
+                        $couponMessage.text(enupalStripeData.coupon.errorMessage);
+                    }
+                    $couponButton.prop('disabled', false);
+                }.bind(this),
+                error: function(xhr, status, err) {
+                    $couponButton.prop('disabled', false);
+                    console.error(xhr, status, err.toString());
+                }.bind(this)
+            });
         },
 
         createCardElement: function(stripe, enupalStripeData, elements, enupalButtonElement) {
@@ -288,6 +385,19 @@ var enupalStripe = {};
             });
         },
 
+        getIsRecurring: function(enupalButtonElement, enupalStripeData) {
+            var isRecurring = false;
+            if (enupalStripeData.amountType == 1) {
+                isRecurring = enupalButtonElement.find('[name="enupalStripe[recurringToggle]"]').is(":checked");
+            }
+
+            if (enupalStripeData.enableSubscriptions) {
+                isRecurring = true;
+            }
+
+            return isRecurring;
+        },
+
         getFinalAmount: function(enupalButtonElement, enupalStripeData) {
             // We always return a default amount
             var finalAmount = enupalStripeData.stripe.amount;
@@ -327,7 +437,7 @@ var enupalStripe = {};
                     // Custom plan
                     var customPlanAmountId = null;
                     if (enupalStripeData.subscriptionStyle == 'radio') {
-                        customPlanAmountId = $('input[name="enupalStripe[enupalMultiPlan]"]:checked').val();
+                        customPlanAmountId = enupalButtonElement.find('input[name="enupalStripe[enupalMultiPlan]"]:checked').val();
                     } else {
                         customPlanAmountId = enupalButtonElement.find('[name="enupalStripe[enupalMultiPlan]"]').val();
                     }
@@ -362,7 +472,29 @@ var enupalStripe = {};
                 enupalButtonElement.find('[name="tax-amount-label"]').empty().append(taxLabel);
             }
 
-            return parseFloat(finalAmount) + parseFloat(fee);
+            var finalTotalAmount = parseFloat(finalAmount) + parseFloat(fee);
+
+            return finalTotalAmount;
+        },
+
+        updatesTotalLabelOnChoosesPlan(enupalButtonElement, enupalStripeData)
+        {
+            var that = this;
+            if (enupalStripeData.subscriptionType !== 0){
+                if (enupalStripeData.subscriptionStyle == 'radio') {
+                    var radio = enupalButtonElement.find('input[name="enupalStripe[enupalMultiPlan]"]');
+                    $(radio).change(function(){
+                        if ($(this).is(':checked')) {
+                            that.handleRemoveCoupon(enupalButtonElement, enupalStripeData, that);
+                        }
+                     });
+                } else {
+                    var dropdown =  enupalButtonElement.find('[name="enupalStripe[enupalMultiPlan]"]');
+                    $(dropdown).on('change', function() {
+                        that.handleRemoveCoupon(enupalButtonElement, enupalStripeData, that);
+                    });
+                }
+            }
         },
 
         convertToCents: function(amount, currency) {